refactor(dashboard): separate Group frame rendering
